refactor: simplify import paths and restructure component files
