+++ src/main.js
@@ -1,5 +1,5 @@
 import { createApp } from "vue";
-import App from "@/organisms/App/App.vue";
+import App from "@/organisms/App.vue";
 import router from "@/router";
 import store from "@/store";
 import "@/styles/main.scss";
